solved mypy error in base.py (#1285)

+++ pymc_marketing/mmm/base.py
@@ -549,7 +549,7 @@ class MMMModelBuilder(ModelBuilder):
             hdi_list = [0.94, 0.5]
 
         if hdi_list and not add_gradient:
-            alpha_list = np.linspace(0.2, 0.4, len(hdi_list))
+            alpha_list = np.linspace(0.2, 0.4, len(hdi_list), dtype=float)
             for hdi_prob, alpha in zip(hdi_list, alpha_list, strict=True):
                 ax = self._add_hdi_to_plot(
                     ax=ax,
